client/requirements.txt
Django==2.1.7
grpcio==1.19.0
mysqlclient==1.3.12
djangorestframework==3.9.2
grpcio-tools